--- a/benchmark_test.go
+++ b/benchmark_test.go
@@ -22,13 +22,14 @@ func BenchmarkContinue(b *testing.B) {
 }
 
 func BenchmarkDo(b *testing.B) {
+	err := errors.New("error")
 	policy := &Policy{
 		MaxCount: 5,
 	}
 	for i := 0; i < b.N; i++ {
 		policy.Do(context.Background(), func() error {
 			dummyFunc()
-			return errors.New("error")
+			return err
 		})
 	}
 }