before:

```
BenchmarkContinue-4   	   70143	     14485 ns/op	     832 B/op	      12 allocs/op
BenchmarkDo-4   	   62914	     17216 ns/op	     992 B/op	      22 allocs/op
```

after:

```
BenchmarkContinue-4   	   78696	     14644 ns/op	     208 B/op	       3 allocs/op
BenchmarkDo-4   	   63008	     17721 ns/op	     368 B/op	      13 allocs/op
```