@shogo82148 2019-09-28
```
BenchmarkDo-4   	   72289	     15469 ns/op	     288 B/op	       8 allocs/op
```