@shogo82148 2019-09-28
Go's escape analysis works file.

```
BenchmarkContinue-4   	 1249548	       890 ns/op	       0 B/op	       0 allocs/op
BenchmarkDo-4   	  433174	      2644 ns/op	      80 B/op	       5 allocs/op
```